requirements.txt
# Core
pandas>=2.0.0
openpyxl>=3.1.0
xlsxwriter>=3.1.0
requests>=2.31.0
python-dotenv>=1.0.0

# Streamlit
streamlit>=1.28.0

# Celery + Redis
celery>=5.3.0
redis>=5.0.0

# Utilitários
Pillow>=10.0.0

# Google Drive
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.100.0